Handle play command (#152)
* Create playing.js

* Update play.js

* Update playlist.js

* Update search.js

* Update playing.js
